+++ src/Request/ClearingTypes.php
@@ -12,4 +12,5 @@ class ClearingTypes
     const ONLINE_BANK_TRANSFER = 'sb';
     const WALLET = 'wlt';
     const FINANCING = 'fnc';
+    const POV = 'pov';
 }